Move Engine.ExpandRequest into SourceHost; lift tests
