+++ source_test.go
@@ -8,197 +8,7 @@ import (
 	"github.com/overmindtech/sdp-go"
 )
 
-func TestFilterSources(t *testing.T) {
-	e, err := NewEngine()
-	if err != nil {
-		t.Fatalf("Error initializing Engine: %v", err)
-	}
-
-	e.AddSources(
-		&TestSource{
-			ReturnScopes: []string{"test"},
-			ReturnType:   "person",
-		},
-		&TestSource{
-			ReturnScopes: []string{"test"},
-			ReturnType:   "fish",
-		},
-		&TestSource{
-			ReturnScopes: []string{sdp.WILDCARD},
-			ReturnType:   "person",
-		},
-		&TestSource{
-			ReturnScopes: []string{
-				"testA",
-				"testB",
-			},
-			ReturnType: "chair",
-		},
-		&TestSource{
-			ReturnScopes: []string{"test"},
-			ReturnType:   "hidden_person",
-			IsHidden:     true,
-		},
-	)
-
-	t.Run("Right type wrong scope", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  "person",
-			Scope: "wrong",
-		}
-
-		ee := ExpectExpand{
-			NumRequests: 1,
-			NumSources:  1,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-	})
-
-	t.Run("Right scope wrong type", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  "wrong",
-			Scope: "test",
-		}
-
-		ee := ExpectExpand{
-			NumRequests: 0,
-			NumSources:  0,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-	})
-
-	t.Run("Right both", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  "person",
-			Scope: "test",
-		}
-
-		ee := ExpectExpand{
-			NumRequests: 1,
-			NumSources:  2,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-	})
-
-	t.Run("Multi-scope", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  "chair",
-			Scope: "testB",
-		}
-
-		ee := ExpectExpand{
-			NumRequests: 1,
-			NumSources:  1,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-	})
-
-	t.Run("Wildcard scope", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  "person",
-			Scope: sdp.WILDCARD,
-		}
-
-		ee := ExpectExpand{
-			NumRequests: 2,
-			NumSources:  2,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-
-		req = sdp.ItemRequest{
-			Type:  "chair",
-			Scope: sdp.WILDCARD,
-		}
-
-		ee = ExpectExpand{
-			NumRequests: 2,
-			NumSources:  2,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-	})
-
-	t.Run("Wildcard type", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  sdp.WILDCARD,
-			Scope: "test",
-		}
-
-		ee := ExpectExpand{
-			NumRequests: 2,
-			NumSources:  3,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-	})
-
-	t.Run("Wildcard both", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  sdp.WILDCARD,
-			Scope: sdp.WILDCARD,
-		}
-
-		ee := ExpectExpand{
-			NumRequests: 5,
-			NumSources:  5,
-		}
-
-		ee.Validate(t, e.ExpandRequest(&req))
-	})
-
-	t.Run("Listing hidden source with wildcard scope", func(t *testing.T) {
-		req := sdp.ItemRequest{
-			Type:  "hidden_person",
-			Scope: sdp.WILDCARD,
-		}
-		if x := len(e.ExpandRequest(&req)); x != 0 {
-			t.Errorf("expected to find 0 sources, found %v", x)
-		}
-
-		req = sdp.ItemRequest{
-			Type:  "hidden_person",
-			Scope: "test",
-		}
-		if x := len(e.ExpandRequest(&req)); x != 1 {
-			t.Errorf("expected to find 1 sources, found %v", x)
-		}
-	})
-}
-
-type ExpectExpand struct {
-	NumRequests int
-
-	// Note that this is not the number of unique sources, but teh number of
-	// sources total. So if a source would be hit twice this will be 2
-	NumSources int
-}
-
-func (e *ExpectExpand) Validate(t *testing.T, m map[*sdp.ItemRequest][]Source) {
-	t.Helper()
-
-	numSources := 0
-	numRequests := 0
-
-	for _, v := range m {
-		numRequests++
-		numSources = numSources + len(v)
-	}
-
-	if e.NumRequests != numRequests {
-		t.Errorf("Expected %v requests, got %v", e.NumRequests, numRequests)
-	}
-
-	if e.NumSources != numSources {
-		t.Errorf("Expected %v sources, got %v", e.NumSources, numSources)
-	}
-}
-
-func TestSourceAdd(t *testing.T) {
+func TestEngineAddSources(t *testing.T) {
 	e, err := NewEngine()
 	if err != nil {
 		t.Fatalf("Error initializing Engine: %v", err)
@@ -266,7 +76,7 @@ func TestGet(t *testing.T) {
 		})
 
 		if err == nil {
-			t.Error("execpected all sources failed")
+			t.Error("expected all sources failed")
 		}
 
 		if len(errs) == 1 {
@@ -736,7 +546,6 @@ func TestListSearchCaching(t *testing.T) {
 			t.Errorf("Exected 2 Search calls, got %v", l)
 		}
 	})
-
 }
 
 func TestSearchGetCaching(t *testing.T) {
